Enhanced the feature of collapsing sidebar
