+++ layouts/dashboard.vue
@@ -1,12 +1,12 @@
 <template>
   <div class="wrapper wrapper--dashboard">
-    <dash-sidebar :collapsed="!collapsedSidebar" :page-move-pending="isChildPending" @toggleSidebar="collapsed" />
+    <dash-sidebar :collapsed="collapsedSidebar" @toggleSidebar="collapseSidebar" />
     <div :class="['content-page position-relative', { collapsed: !collapsedSidebar }]">
       <LoadingWrapper :loading="isChildPending" />
       <transition name="dashboard-fade" mode="out-in" appear>
         <div v-show="!isChildPending" class="content">
           <div class="dash-nav-virtual-divider" />
-          <dash-nav class="mb-0 mb-md-5" :collapsed="!collapsedSidebar" @toggleSidebar="collapsed" />
+          <dash-nav class="mb-0 mb-md-5" :collapsed="!collapsedSidebar" @toggleSidebar="collapseSidebar" />
           <Nuxt id="content-body" class="content-body" />
         </div>
       </transition>
@@ -15,16 +15,34 @@
 </template>
 
 <script>
+import checkResolution from '~/mixins/checkResolution'
+
 export default {
   name: 'Dashboard',
   components: {
     LoadingWrapper: () => { return import('~/components/loadingWrapper') }
   },
+  mixins: [
+    checkResolution
+  ],
   data: () => ({
-    collapsedSidebar: false,
+    collapsedSidebar: true,
     isChildPending: true
   }),
-  watch: {},
+  watch: {
+    isMobile (aft, prev) {
+      if (!this.isChildPending && prev !== null) {
+        this.collapseSidebar(aft)
+      }
+    },
+    isChildPending (aft, prev) {
+      if (this.isMobile && !this.collapsedSidebar && (prev && !aft)) {
+        setTimeout(() => {
+          this.collapseSidebar(true)
+        }, 1000)
+      }
+    }
+  },
   created () {
     this.$nuxt.$on('pageLoading', (status) => {
       const _status = typeof status === 'undefined' ? false : status
@@ -32,12 +50,24 @@ export default {
     })
   },
   mounted () {
-    this.collapsedSidebar = window.innerWidth < 1024
+    this.collapseSidebar(window.innerWidth <= 1024)
+    window.addEventListener('mouseup', this.handleSidebar)
   },
-  beforeDestroy () {},
+  beforeDestroy () {
+    window.removeEventListener('mouseup', this.handleSidebar)
+  },
   methods: {
-    collapsed (state = !this.collapsedSidebar) {
+    collapseSidebar (state = !this.collapsedSidebar) {
       this.collapsedSidebar = state
+    },
+    handleSidebar (event) {
+      if (this.isMobile && !this.collapsed) {
+        const headerChildEle = event.target.closest('#dash-nav')
+        const sidebarChildEle = event.target.closest('#dash-sidebar')
+        if (!this.isChildPending && !sidebarChildEle && !headerChildEle) {
+          this.collapseSidebar()
+        }
+      }
     }
   }
 }
